services/identity-provider/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
psycopg[binary]==3.1.18
PyJWT==2.9.0
